% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_data_documentation.R
\docType{data}
\name{fish_data}
\alias{fish_data}
\title{Fish stock and harvest data for a hypothetical fisheries model}
\format{
A data frame with 101 rows and 2 variables.
}
\usage{
fish_data
}
\description{
Random data produced using the 'rnorm' function with 100 observations.
Stock data has a mean value of 100,000 (kg)
Harvest data (annual) has a mean value of 10,000 (kg)
}
\keyword{datasets}
