% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_stocks_documentation.R
\docType{data}
\name{fish_stocks}
\alias{fish_stocks}
\title{This is a test}
\format{
An object of class \code{numeric} of length 10000.
}
\usage{
fish_stocks
}
\description{
This is a test
}
\keyword{datasets}
