% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_rates_documentation.R
\docType{data}
\name{harvest_rates}
\alias{harvest_rates}
\title{Test for harvest rates documentation}
\format{
An object of class \code{numeric} of length 100.
}
\usage{
harvest_rates
}
\description{
Test for harvest rates documentation
}
\keyword{datasets}
