% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcount_documentation.R
\docType{data}
\name{newcount}
\alias{newcount}
\title{New count data test}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
newcount
}
\description{
This is simply to make sure you can create a help page for
See examples for
}
\keyword{datasets}
