% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_data_documentation.R
\docType{data}
\name{fish_data}
\alias{fish_data}
\title{Documentation for fish data}
\format{
An object of class \code{data.frame} with 100 rows and 2 columns.
}
\usage{
fish_data
}
\description{
Documentation for fish data
}
\keyword{datasets}
