% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_harvest.R
\name{fish_harvest}
\alias{fish_harvest}
\title{This function is a basic fisheries harvest model. It takes the initial stock biomass, annual harvest rate, and number of years as inputs.}
\usage{
fish_harvest(
  stock_initial,
  annual_harvest,
  growth_rate = 0.5,
  carrying_capacity = stock_initial,
  num_years
)
}
\arguments{
\item{stock_initial}{(kg)}

\item{annual_harvest}{(kg)}

\item{growth_rate}{(kg) default is .5}

\item{carrying_capacity}{= stock_initial}
}
\value{
final stock (kg)

sustainability of harvest (sustainable or unsustainable)

status of stock (healthy, overfished, critically low or fully depleted)
}
\description{
This function is a basic fisheries harvest model. It takes the initial stock biomass, annual harvest rate, and number of years as inputs.
}
