% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_stocks_documentation.R
\docType{data}
\name{fish_stocks}
\alias{fish_stocks}
\title{This is is sample data for fish stocks (kg) for the initial level of a theoretical harvest scenario}
\format{
A vector with 100 values with a mean of 10,000 and standard deviation of (10,000*.15)
}
\usage{
fish_stocks
}
\description{
This is is sample data for fish stocks (kg) for the initial level of a theoretical harvest scenario
}
\keyword{datasets}
