% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_rates_documentation.R
\docType{data}
\name{harvest_rates}
\alias{harvest_rates}
\title{Test for harvest rates documentation}
\format{
A vector with 100 values with a mean of 10,000 and standard deviation of (10,000*.15)
}
\usage{
harvest_rates
}
\description{
Test for harvest rates documentation
}
\keyword{datasets}
